BACKEND/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
pillow==12.1.1
aiofiles==25.1.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic>=2.9.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
reportlab>=4.0.7
weasyprint>=59.0
opencv-python>=4.9.0
numpy>=1.26.0
google-cloud-vision>=3.5.0
